Move project to the humble_downloader subfolder
This prepares the introduction of a setup.py file and makes things neater.
Also fix imports following the renaming, and add some.
