+++ humble_downloader/configuration.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 import argparse
 import os
 import yaml
-import logger
-from config_data import ConfigData
-from humble_api.humble_hash import HumbleHash
+from . import logger
+from .config_data import ConfigData
+from .humble_api.humble_hash import HumbleHash
 
 __author__ = "Brian Schkerke"
 __copyright__ = "Copyright 2016 Brian Schkerke"
