+++ humble_downloader/event_handler.py
@@ -1,8 +1,9 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+
 import sys
-import logger
-from humble_api.events import Events
+from . import logger
+from .humble_api.events import Events
 
 __author__ = "Brian Schkerke"
 __copyright__ = "Copyright 2016 Brian Schkerke"
