+++ humble_downloader/humble_api/exceptions/__init__.py
