+++ humble_downloader/humble_api/exceptions/humble_authentication_exception.py
