+++ humble_downloader/humble_api/exceptions/humble_download_needed_exception.py
