+++ humble_downloader/humble_api/exceptions/humble_exception.py
