+++ humble_downloader/humble_api/exceptions/humble_parse_exception.py
