+++ humble_downloader/humble_api/model/base_model.py
