+++ humble_downloader/humble_api/model/download.py
