+++ humble_downloader/humble_api/model/download_struct.py
