+++ humble_downloader/humble_api/model/product.py
