+++ humble_downloader/humble_api/model/subscription.py
