+++ humble_downloader/logger.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from config_data import ConfigData
+
+from .config_data import ConfigData
 import time
 import sys
